fix(calendar): MultiDayView and MonthView were collapsing to content width
Critical layout regression caught from a screenshot: in week / 3-day /
month views, the day columns were squished into the leftmost ~30% of
the viewport with the right ~70% empty.

Root cause: both `MultiDayView` and `MonthView` rendered as the only
flex child of `<div className="flex flex-1 overflow-hidden">` (a flex
ROW container) with their own outer wrapper of just
`flex h-full flex-col`. Without `flex-1` or `w-full`, a flex-row
child collapses to its content's intrinsic width — which for the
calendar was just the natural width of its day columns at minimum
content size, not the full viewport.

Day view didn't show the bug because its inner Timeline component had
flex-1 set on its content, masking the parent's lack of width.

Fix: add `w-full min-w-0 flex-1` to both view wrappers so they fill
the available row width.

Also decomposed `create-time-block-dialog.tsx` into three files
(`create-time-block-dialog.tsx` orchestrator, plus
`create-dialog/time-block-form.tsx` and `create-dialog/event-form.tsx`)
to keep each file under the 250-line guideline — flagged by the
verifier on the previous commit.
